docs: add fixable notice to rule doc of fixable rules and test for this
