+++ docs/rule/inline-link-to.md
@@ -1,5 +1,7 @@
 # inline-link-to
 
+:wrench: The `--fix` option on the command line can automatically fix some of the problems reported by this rule.
+
 Ember's `link-to` component has both an inline form and a block form. This rule forbids the inline form.
 
 ## Examples
