+++ docs/rule/require-button-type.md
@@ -2,6 +2,8 @@
 
 :white_check_mark: The `extends: 'recommended'` property in a configuration file enables this rule.
 
+:wrench: The `--fix` option on the command line can automatically fix some of the problems reported by this rule.
+
 This rule requires all `<button>` elements to have a valid `type` attribute.
 
 By default, the `type` attribute of `<button>` elements is `submit`. This can
