+++ scripts/update-rules.js
@@ -2,15 +2,12 @@
 
 const fs = require('fs');
 const path = require('path');
-
-const { parse } = require('@babel/parser');
-const { default: traverse } = require('@babel/traverse');
 const prettier = require('prettier');
-
 const rules = require('../lib/rules');
 const { rules: recommendedRules } = require('../lib/config/recommended');
 const { rules: octaneRules } = require('../lib/config/octane');
 const { rules: stylisticRules } = require('../lib/config/stylistic');
+const isRuleFixable = require('../test/helpers/is-rule-fixable');
 
 const prettierConfig = {
   ...require('../.prettierrc.js'),
@@ -26,32 +23,6 @@ const EMOJI_STAR = ':white_check_mark:';
 const EMOJI_OCTANE = ':car:';
 const EMOJI_STYLISTIC = ':dress:';
 const EMOJI_FIXABLE = ':wrench:';
-
-function isRuleFixable(ruleName) {
-  const relativePath = ruleName.startsWith('deprecated-')
-    ? `../lib/rules/deprecations/${ruleName}.js`
-    : `../lib/rules/${ruleName}.js`;
-  const pathRule = path.resolve(__dirname, relativePath);
-  let rule = fs.readFileSync(pathRule, { encoding: 'utf8' });
-
-  let ast = parse(rule, { sourceType: 'module' });
-
-  let isFixable = false;
-
-  traverse(ast, {
-    ObjectProperty(path) {
-      if (
-        path.node.key.type === 'Identifier' &&
-        path.node.key.name === 'isFixable' &&
-        !(path.node.value.type === 'BooleanLiteral' && path.node.value.value === false)
-      ) {
-        isFixable = true;
-      }
-    },
-  });
-
-  return isFixable;
-}
 
 // Generate rule table contents.
 const rulesTableContent = Object.keys(rules)
